test_task/requirements.txt
Django==5.0
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
drf-spectacular==0.26.5
django-cors-headers==4.3.1
gunicorn==21.2.0
psycopg2-binary~=2.0
Pillow==10.1.0
django-environ==0.11.2
openai==1.33.0